Phd 119 (#9)
* Export variables into .env file

* Adding REGISTER_PATH to .env

* Handle differently spoon and spring-boot

* Adding setter for dummy API_URL for test

* Delete .env file from repo

+++ run.sh
@@ -1,21 +1,12 @@
 #!/bin/bash
 
-project_name="spring-boot"
-src_folder_path="/spring-boot-project/spring-boot/"
+export $(grep -v '^#' .env | xargs -d '\n')
 
-#project_name="spoon"
-#src_folder_path=""
+input_repo_path="${REPO_PATH}${PROJECT_NAME}"
+input_source_code="${input_repo_path}${SRC_FOLDER_PATH}/src/main/java/"
 
-#project_name="jabref"
-#src_folder_path=""
-
-repo_path="/home/jerome/Documents/Assistant/Recherche/joular-scripts/test-sentinel-copy/test/"
-
-input_repo_path="${repo_path}${project_name}"
-input_source_code="${input_repo_path}${src_folder_path}/src/main/java/"
-
-output_repo_path="${repo_path}${project_name}_output"
-output_source_code="${output_repo_path}${src_folder_path}/src/main/java/"
+output_repo_path="${REPO_PATH}${PROJECT_NAME}_output"
+output_source_code="${output_repo_path}${SRC_FOLDER_PATH}/src/main/java/"
 
 # Remove old output folder
 echo "Removing ${output_repo_path}..."
@@ -27,26 +18,34 @@ echo "Copying ${input_repo_path} into ${output_repo_path}"
 echo "cp -r ${input_repo_path} ${output_repo_path}"
 cp -r "$input_repo_path" "$output_repo_path"
 
+# Create the .jar of object-instantiation project
 echo "mvn clean package"
 mvn clean package
 
+# Execute the .jar on the analyzed project
 cd ./target || return
 echo "java -jar object-instantiation-1.0-SNAPSHOT-jar-with-dependencies.jar ${input_source_code} ${output_source_code}"
 java -jar object-instantiation-1.0-SNAPSHOT-jar-with-dependencies.jar "$input_source_code" "$output_source_code" "$input_repo_path"
 
-# Copy be.unamur.snail.register.SendUtils to the output folder
-register_path="/home/jerome/Documents/Assistant/Recherche/joular-scripts/object-instantiation/src/main/java/be/unamur/snail/register/."
+# Copy be.unamur.snail.register package to the output folder
 echo "mkdir -p ${output_source_code}/be/unamur/snail/register"
 mkdir -p "${output_source_code}/be/unamur/snail/register"
-echo "cp -r ${register_path} ${output_source_code}/be/unamur/snail/register"
-cp -r "$register_path" "${output_source_code}/be/unamur/snail/register/"
+echo "cp -r ${REGISTER_PATH} ${output_source_code}/be/unamur/snail/register"
+cp -r "$REGISTER_PATH" "${output_source_code}/be/unamur/snail/register/"
 
-# Execute spring-boot tests with the transformation
+# Execute tests of the analyzed project within the transformed code
 cd "$output_repo_path" || return
-rm -rf .gradle/
-export JAVA_HOME=/usr/lib/jvm/java-19-openjdk-amd64
-echo "./gradlew clean spring-boot-project:spring-boot:test"
-./gradlew clean spring-boot-project:spring-boot:test --rerun-tasks
 
-#export JAVA_HOME=/usr/lib/jvm/java-17-openjdk-amd64
-#mvn clean test
+if [ "$PROJECT_NAME" == "spring-boot" ]; then
+  export JAVA_HOME=/usr/lib/jvm/java-19-openjdk-amd64
+  rm -rf .gradle/
+  echo "./gradlew clean spring-boot-project:spring-boot:test"
+  ./gradlew clean spring-boot-project:spring-boot:test --rerun-tasks
+fi
+
+if [ "$PROJECT_NAME" == "spoon" ]; then
+  export JAVA_HOME=/usr/lib/jvm/java-17-openjdk-amd64
+  mvn clean test
+fi
+
+# TODO handle other projects execution
